requirements.txt
streamlit==1.26.0
torch==2.1.1
torchvision==0.16.1
transformers==4.38.0
huggingface_hub==0.20.3
Pillow==9.5.0
matplotlib==3.8.0

